Change user to client, add receivers, remove debug code, add support for HTTP/2 & tls
