+++ .gitignore
@@ -22,3 +22,8 @@ _testmain.go
 *.exe
 *.test
 *.prof
+
+# Keys
+
+localhost.cert
+localhost.key
