+++ util/authentication.go
@@ -1,0 +1,7 @@
+package util
+
+// CheckLoginData Check set of login data, and return a bool.
+func CheckLoginData(channel string, ip string, authString string) bool {
+	// TODO
+	return true
+}
